requirements.txt
alembic==1.14.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asttokens==3.0.0
bcrypt==4.2.1
black==25.1.0
certifi==2025.8.3
cffi==1.17.1
click==8.1.8
colorama==0.4.6
coverage==7.10.5
cryptography==44.0.0
DateTime==5.5
decorator==5.1.1
executing==2.2.0
flake8==7.3.0
iniconfig==2.1.0
ipython==8.31.0
jedi==0.19.2
Mako==1.3.9
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==25.0
parso==0.8.4
pathspec==0.12.1
platformdirs==4.4.0
pluggy==1.6.0
prompt_toolkit==3.0.50
pure_eval==0.2.3
pycodestyle==2.14.0
pycparser==2.22
pyflakes==3.4.0
Pygments==2.19.1
PyJWT==2.10.1
PyMySQL==1.1.1
pytest==8.4.1
pytest-cov==6.2.1
pytest-mock==3.14.1
pytz==2025.1
sentry-sdk==2.35.0
setuptools==75.8.2
SQLAlchemy==2.0.36
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.5.0
wcwidth==0.2.13
zope.interface==7.2
